requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.16.1
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0
coverage==7.3.2
black==23.10.0
flake8==6.1.0
isort==5.12.0